+++ src/components/readBoard/MoreBoxButton.tsx
@@ -32,6 +32,5 @@ const Wrapper = styled.div`
 `;
 
 const ThreeDotsImage = styled(Image)`
-  width: 15px;
   cursor: pointer;
 `;
